Create new package for oracles (#1524)
