+++ .github/workflows/ci.yml
@@ -254,3 +254,17 @@ jobs:
           path: pkg/
       - name: Test
         run: yarn workspace @balancer-labs/v3-pool-cow test:forge
+
+  test-forge-oracles:
+    runs-on: ubuntu-latest
+    needs: lint-and-build
+    steps:
+      - uses: actions/checkout@v5
+      - name: Set up environment
+        uses: ./.github/actions/setup
+      - uses: actions/download-artifact@v4
+        with:
+          name: built-contracts
+          path: pkg/
+      - name: Test
+        run: yarn workspace @balancer-labs/v3-oracles test:forge
