+++ pkg/interfaces/contracts/oracles/IWeightedLPOracle.sol
