+++ pkg/oracles/.solhintignore
@@ -1,0 +1,1 @@
+contracts/test/
