+++ pkg/oracles/contracts/test/StableLPOracleMock.sol
