+++ pkg/oracles/coverage.sh
@@ -1,0 +1,1 @@
+../../coverage.sh
