+++ pkg/oracles/hardhat.config.ts
@@ -1,0 +1,15 @@
+import '@nomicfoundation/hardhat-ethers';
+import '@nomicfoundation/hardhat-toolbox';
+import '@typechain/hardhat';
+
+import 'hardhat-ignore-warnings';
+import 'hardhat-gas-reporter';
+
+import { hardhatBaseConfig } from '@balancer-labs/v3-common';
+
+export default {
+  solidity: {
+    compilers: hardhatBaseConfig.compilers,
+  },
+  warnings: hardhatBaseConfig.warnings,
+};
