+++ pkg/oracles/package.json
@@ -1,0 +1,54 @@
+{
+  "name": "@balancer-labs/v3-oracles",
+  "version": "0.1.0",
+  "description": "Balancer V3 Oracles",
+  "license": "GPL-3.0-only",
+  "homepage": "https://github.com/balancer-labs/balancer-v3-monorepo/tree/master/pkg/oracles#readme",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/balancer-labs/balancer-v3-monorepo.git",
+    "directory": "pkg/oracles"
+  },
+  "bugs": {
+    "url": "https://github.com/balancer-labs/balancer-v3-monorepo/issues"
+  },
+  "files": [
+    "contracts/**/*.sol",
+    "!contracts/test/**/*.sol"
+  ],
+  "scripts": {
+    "build": "yarn compile && rm -rf artifacts/build-info",
+    "compile": "hardhat compile",
+    "compile:watch": "nodemon --ext sol --exec yarn compile",
+    "lint": "yarn lint:solidity && yarn lint:typescript",
+    "lint:solidity": "npx prettier --check --plugin=prettier-plugin-solidity 'contracts/**/*.sol' ''test/**/*.sol'' && npx solhint 'contracts/**/*.sol'",
+    "lint:typescript": "NODE_NO_WARNINGS=1 eslint . --ext .ts --ignore-path ../../.eslintignore  --max-warnings 0",
+    "prettier": "npx prettier --write --plugin=prettier-plugin-solidity 'contracts/**/*.sol' 'test/**/*.sol'",
+    "test": "yarn test:hardhat && yarn test:forge",
+    "test:hardhat": "hardhat test",
+    "test:forge": "yarn build && REUSING_HARDHAT_ARTIFACTS=true forge test --ffi -vvv",
+    "test:stress": "FOUNDRY_PROFILE=intense forge test --ffi -vvv",
+    "coverage": "./coverage.sh forge",
+    "coverage:hardhat": "./coverage.sh hardhat",
+    "coverage:all": "./coverage.sh all",
+    "gas": "REPORT_GAS=true hardhat test",
+    "test:watch": "nodemon --ext js,ts --watch test --watch lib --exec 'clear && yarn test --no-compile'",
+    "slither": "yarn compile --force && bash -c 'source ../../slither/bin/activate && slither --compile-force-framework hardhat --ignore-compile . --config-file ../../.slither.config.json'",
+    "slither:triage": "yarn compile && bash -c 'source ../../slither/bin/activate && slither --compile-force-framework hardhat --ignore-compile . --config-file ../../.slither.config.json --triage-mode'"
+  },
+  "dependencies": {
+    "@balancer-labs/v3-interfaces": "workspace:*"
+  },
+  "devDependencies": {
+    "@balancer-labs/solidity-toolbox": "workspace:*",
+    "@balancer-labs/v3-solidity-utils": "workspace:*",
+    "@typescript-eslint/eslint-plugin": "^5.41.0",
+    "@typescript-eslint/parser": "^5.41.0",
+    "eslint": "^8.26.0",
+    "eslint-plugin-mocha-no-only": "^1.1.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "hardhat": "^2.14.0",
+    "mocha": "^10.1.0",
+    "solhint": "^3.4.1"
+  }
+}
