+++ pkg/oracles/test/gas/.hardhat-snapshots/EclpLPOracle - 2 tokens
