+++ pkg/oracles/test/gas/.hardhat-snapshots/StableLPOracle - 2 tokens
