+++ pkg/oracles/test/gas/.hardhat-snapshots/StableLPOracle - 3 tokens
