+++ pkg/oracles/test/gas/.hardhat-snapshots/StableLPOracle - 4 tokens
