+++ pkg/oracles/test/gas/.hardhat-snapshots/WeightedLPOracle - 2 tokens
