+++ pkg/oracles/test/gas/.hardhat-snapshots/WeightedLPOracle - 3 tokens
