+++ pkg/oracles/test/gas/.hardhat-snapshots/WeightedLPOracle - 4 tokens
