+++ pkg/oracles/test/gas/.hardhat-snapshots/WeightedLPOracle - 5 tokens
