+++ pkg/oracles/test/gas/.hardhat-snapshots/WeightedLPOracle - 6 tokens
